requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.4
Django==2.2.13
django-jsonfield==1.4.1
flake8==3.8.4
idna==2.10
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
